+++ core/src/main/java/org/dkf/jed2k/Settings.java
@@ -6,8 +6,8 @@ public class Settings {
     public final Hash userAgent = new Hash(Hash.LIBED2K);
     public String modName = new String("jed2k");
     public String clientName = new String("jed2k");
-    public short listenPort = 4661;
-    public short udpPort = 4662;
+    public int listenPort = 4661;
+    public int udpPort = 4662;
     public int version = 0x3c;
     public int modMajor = 0;
     public int modMinor = 0;
